+++ packages/pluggableWidgets/video-player-web/typings/VideoPlayerProps.d.ts
@@ -15,8 +15,8 @@ export interface VideoPlayerContainerProps {
     class: string;
     style?: CSSProperties;
     tabIndex?: number;
-    urlExpression: DynamicValue<string>;
-    posterExpression?: DynamicValue<string>;
+    videoUrl: DynamicValue<string>;
+    posterUrl?: DynamicValue<string>;
     autoStart: boolean;
     showControls: boolean;
     muted: boolean;
@@ -30,8 +30,8 @@ export interface VideoPlayerContainerProps {
 export interface VideoPlayerPreviewProps {
     class: string;
     style: string;
-    urlExpression: string;
-    posterExpression: string;
+    videoUrl: string;
+    posterUrl: string;
     autoStart: boolean;
     showControls: boolean;
     muted: boolean;